separate a temp effect manager
